// EmployeeDirectoryFrontEnd/src/app/shared/models/Employee.ts
export interface Employee {
    id: number;
    employeeId: string;
    firstName: string;
    lastName: string;
    dob: string;
    emailID: string;
    phoneNo: string;
    joinDate: string;
    location: string;
    jobTitle: string;
    department: string;
    manager: string;
    project: string;
    profilePicture: string;
}